refactor: player-rating-rotator players-rating-rotator오타수정
