+++ src/domains/players/players-rating-rotator/api/players-rating-rotator-api.ts
@@ -1,4 +1,5 @@
 import { supabase } from "@shared/api/config/supabaseClient";
+
 // axios가 필요없다. why? supabase rls설정 rpc를 통하여 권한체크 가능함
 
 export interface IRotatePlayerStatAccumulated {
@@ -6,7 +7,7 @@ export interface IRotatePlayerStatAccumulated {
   overall_avg_rating: number | null;
 }
 
-export const getPlayerRatingRotatorAcc = async () => {
+export const getPlayersRatingRotatorAcc = async () => {
   const { data, error } = await supabase.rpc("get_player_rating_rotator_acc");
 
   if (error) {
