+++ src/domains/players/players-rating-rotator/components/error/players-rating-rotator-error-fallback.tsx
@@ -1,0 +1,29 @@
+import PlayersRatingRotatorWrapper from "../wrapper/players-rating-rotator-wrapper";
+
+/**
+ * 작성자: KYD
+ * 기능:
+ * 프로세스 설명: 프로세스 복잡시 노션링크 첨부권장
+ */
+
+const PlayerRatingRotatorErrorFallback = () => {
+  //SECTION HOOK호출 영역
+
+  //!SECTION HOOK호출 영역
+
+  //SECTION 상태값 영역
+
+  //!SECTION 상태값 영역
+
+  //SECTION 메서드 영역
+
+  //!SECTION 메서드 영역
+
+  return (
+    <PlayersRatingRotatorWrapper>
+      <div className="flex w-full items-center justify-center">보루센 평점을 일시적으로 가져오지 못했습니다.</div>
+    </PlayersRatingRotatorWrapper>
+  );
+};
+
+export default PlayerRatingRotatorErrorFallback;
