--- src/domains/players/players-rating-rotator/components/player-rating-rotator/skeleton/player-rating-rotator-skeleton.tsx
@@ -1,29 +1,0 @@
-/**
- * 작성자: KYD
- * 기능:
- * 프로세스 설명: 프로세스 복잡시 노션링크 첨부권장
- */
-
-import PlayerRatingRotatorWrapper from "../wrapper/player-rating-rotator-wrapper";
-
-const PlayerRatingRotatorSkeleton = () => {
-  //SECTION HOOK호출 영역
-
-  //!SECTION HOOK호출 영역
-
-  //SECTION 상태값 영역
-
-  //!SECTION 상태값 영역
-
-  //SECTION 메서드 영역
-
-  //!SECTION 메서드 영역
-
-  return (
-    <PlayerRatingRotatorWrapper>
-      <div className="flex w-full items-center justify-center">보루센 평점 가져오는중...</div>
-    </PlayerRatingRotatorWrapper>
-  );
-};
-
-export default PlayerRatingRotatorSkeleton;
