+++ src/domains/players/players-rating-rotator/components/players-rating-rotator.tsx
@@ -3,16 +3,18 @@
  * 기능: 유저 이목을 끌기 위한 장치로 선수 누적 평점 주식처럼 표시
  * 프로세스 설명: SUPABASE RPC함수 - REACT QUERY 래핑 - 조회
  */
-import { useGetPlayerRatingRotatorAcc } from "@players/players-rating-rotator/api/react-query-api/use-get-player-rating-rotator-acc";
-import gsap from "gsap";
-import { useGSAP } from "@gsap/react";
 import { useRef } from "react";
-import type { IRotatePlayerStatAccumulated } from "@players/players-rating-rotator/api/player-rating-rotator-api";
-import PlayerRatingRotatorWrapper from "./wrapper/player-rating-rotator-wrapper";
 
-const PlayerRatingRotator = () => {
+import PlayersRatingRotatorWrapper from "./wrapper/players-rating-rotator-wrapper";
+import { useGSAP } from "@gsap/react";
+import gsap from "gsap";
+
+import type { IRotatePlayerStatAccumulated } from "@players/players-rating-rotator/api/players-rating-rotator-api";
+import { useGetPlayersRatingRotatorAcc } from "@players/players-rating-rotator/api/react-query-api/use-get-players-rating-rotator-acc";
+
+const PlayersRatingRotator = () => {
   //SECTION HOOK호출 영역
-  const fetchPlayerData = useGetPlayerRatingRotatorAcc();
+  const fetchPlayerData = useGetPlayersRatingRotatorAcc();
   const containerRef = useRef<HTMLDivElement>(null);
   //!SECTION HOOK호출 영역
 
@@ -34,24 +36,24 @@ const PlayerRatingRotator = () => {
   );
 
   return (
-    <PlayerRatingRotatorWrapper>
+    <PlayersRatingRotatorWrapper>
       <div ref={containerRef} className="flex w-[max-content] gap-4">
         {/* 첫 번째 세트 */}
         {fetchPlayerData.map((item) => (
-          <PlayerRatingItem key={`first-${item.korean_name}`} {...item} />
+          <PlayersRatingItem key={`first-${item.korean_name}`} {...item} />
         ))}
         {/* 두 번째 세트 (무한 회전용) */}
         {fetchPlayerData.map((item) => (
-          <PlayerRatingItem key={`second-${item.korean_name}`} {...item} />
+          <PlayersRatingItem key={`second-${item.korean_name}`} {...item} />
         ))}
       </div>
-    </PlayerRatingRotatorWrapper>
+    </PlayersRatingRotatorWrapper>
   );
 };
 
-export default PlayerRatingRotator;
+export default PlayersRatingRotator;
 
-const PlayerRatingItem = ({ korean_name, overall_avg_rating }: IRotatePlayerStatAccumulated) => {
+const PlayersRatingItem = ({ korean_name, overall_avg_rating }: IRotatePlayerStatAccumulated) => {
   const getRatingColor = (rating: number | null) => {
     if (!rating) return "text-rating-blue";
     return rating > 7 ? "text-rating-red" : "text-rating-blue";
