+++ src/domains/players/players-rating-rotator/components/skeleton/players-rating-rotator-skeleton.tsx
@@ -1,12 +1,11 @@
-import PlayerRatingRotatorWrapper from "../wrapper/player-rating-rotator-wrapper";
-
 /**
  * 작성자: KYD
  * 기능:
  * 프로세스 설명: 프로세스 복잡시 노션링크 첨부권장
  */
+import PlayerRatingRotatorWrapper from "../wrapper/players-rating-rotator-wrapper";
 
-const PlayerRatingRotatorErrorFallback = () => {
+const PlayerRatingRotatorSkeleton = () => {
   //SECTION HOOK호출 영역
 
   //!SECTION HOOK호출 영역
@@ -21,9 +20,11 @@ const PlayerRatingRotatorErrorFallback = () => {
 
   return (
     <PlayerRatingRotatorWrapper>
-      <div className="flex w-full items-center justify-center">보루센 평점을 일시적으로 가져오지 못했습니다.</div>
+      <div data-testid="players-rating-rotator-skeleton" className="flex w-full items-center justify-center">
+        보루센 평점 가져오는중...
+      </div>
     </PlayerRatingRotatorWrapper>
   );
 };
 
-export default PlayerRatingRotatorErrorFallback;
+export default PlayerRatingRotatorSkeleton;
