+++ src/domains/players/players-rating-rotator/components/wrapper/players-rating-rotator-wrapper.tsx
@@ -3,7 +3,7 @@
  * 기능: 에러,스켈레톤,구현 컴포넌트를 감싸는 컨테이너
  * 프로세스 설명:
  */
-const PlayerRatingRotatorWrapper = ({ children }: { children: React.ReactNode }) => {
+const PlayersRatingRotatorWrapper = ({ children }: { children: React.ReactNode }) => {
   return <div className="relative w-[clamp(320px,100vw,500px)] overflow-hidden p-4 text-lg text-white">{children}</div>;
 };
-export default PlayerRatingRotatorWrapper;
+export default PlayersRatingRotatorWrapper;
