+++ src/shared/mocks/handlers/handlers.ts
@@ -1,5 +1,6 @@
 import { AnimalHandlers } from "@animals/mocks/animal-handlers";
-import { PlayerRatingRotatorHandlers } from "@players/players-rating-rotator/mocks/player-rating-rotator-handlers";
-import { PlayersDBHandlers } from "@players/players-db/mocks/players-db-handler";
 
-export const handlers = [...AnimalHandlers, ...PlayerRatingRotatorHandlers, ...PlayersDBHandlers];
+import { PlayersDBHandlers } from "@players/players-db/mocks/players-db-handler";
+import { PlayersRatingRotatorHandlers } from "@players/players-rating-rotator/mocks/player-rating-rotator-handlers";
+
+export const handlers = [...AnimalHandlers, ...PlayersRatingRotatorHandlers, ...PlayersDBHandlers];
